Refactored etest code: split up test cases into source files.

+++ M3/M3StopWatch.h
@@ -5,7 +5,7 @@
 
 -(void)startWatch;
 -(uint64_t)nanoSeconds;
--(uint64_t)milliSeconds;
+-(int)milliSeconds;
 -(double)seconds;
 
 @end
